Fix up pad_check_dup entry in perlintern
When the pad API was added, the special

   m|pad_check_dup|SV *name|U32 flags|const HV *ourstash

sequence was added to pad.c, but that is unnecessary as it is listed
in embed.fnc.  Also it is not correct, as the name of the function is
in the return value field.

So this restore it back to the simple ‘=for apidoc pad_check_dup’.

The description of the function was in plain text like this:

Check for duplicate declarations: report any of:
     * a my in the current scope with the same name;
     * an our (anywhere in the pad) with the same name and the same stash
       as C<ourstash>
C<is_our> indicates that the name to check is an 'our' declaration

which gets rewrapped in rendered pod.  So this patch changes it to use
a verbatim block (as autodoc.pl doesn’t seem to like pod lists).