[perl #94390] Optimised numeric sort should warn for nan
In this case:

    sort { $a <=> $b } ...

the sort block is optimised away and implemented in C.

That C implementation did not take into account that $a or $b might be
nan, and therefore, without optimisation, would return undef, result-
ing in a warning.

The optimisation is supposed to be just that, and not change
behaviour.