In bisect-runner.pl's synopsis, the test program must be outside the cwd.
Karl notes that the previous version, using C<test_prog.pl>, wrongly
suggests that a test program can be in the git checkout used for
bisecting. This won't work for an untracked file, because bisect.pl's first
sanity check will spot it and refuse to run. For a tracked file (such as an
existing test script in t), things may be far more confusing, as
bisect-runner.pl will end up running the current version for the revision
tested, instead of the version for the revision checked out at start time.