bisect-runner.pl now runs the testcase for targets config.{sh,h}
Previously for these two targets it assumed --test-build if a --match
wasn't supplied, and never ran a test case if one was supplied. Now
--test-build must be specified explicitly, otherwise the test case will be
run. For example, this makes it easy to bisect using a testcase which greps
config.sh or config.h. (Of course, one can do roughly this with the --match
option, but this will match against all generated files, which may generate
false positives.)