bisect-runner.pl now builds test_prep on OpenBSD back to 5.002
The historical OpenBSD hints file needs tweaking for compiler and linker
flags, and needs to be provided for revisions before it was added to the
source tree. perl.h and pp_sys.c need patching with the current (i.e. post
1996) #ifdef forest for [gs]etpgrp() variants. perl.h needs to include
<unistd.h> on OpenBSD, else POSIX.xs won't build.

OpenBSD also requires all the parallel Makefile fixes, as its make builds
targets in reverse lexical order, which reveals a lot of assumptions about
build order. (Such as Cwd nearly always being built in time, because it
sorts lexically far ahead of other XS modules dependant on it.)