Dynamically create and delete testdir/test.lib/
Due to the '.' in 'test.lib's name, t/filenames.t fails (because '.' is the path
separator on VMS). To fix this, I create and delete the directory and its
contents during each test that needs to access the directory. Said tests will be
skipped if the directory cannot be created (e.g., on VMS).