[perl #96326] *{$io} should not be semi-defined
gv_efullname4 produces undef if the GV points to no stash, instead of
using __ANON__, as it does when the stash has no name.

Instead of going through hoops to try and work around it elsewhere, fix
gv_efullname4.

This means that

   $x = *$io;
   $x .= "whate’er";

no longer produces an uninitialized warning.  (The warning was rather
strange, as defined() returned true.)

This commit also gives the glob the name $__ANONIO__ (yes, with a dol-
lar sign).  It may seem a little strange, but there is precedent in
other autovivified globs, such as those open() produces when it cannot
determine the variable name (e.g, open $t->{fh}).